requirements.txt
torch
tochvision
pillow
numpy
skimage
matplotlib